refactor: move Stats into Ledger
