+++ bitswap/src/lib.rs
@@ -10,10 +10,10 @@ mod prefix;
 mod protocol;
 mod strategy;
 
-pub use self::behaviour::{Bitswap, Stats};
+pub use self::behaviour::Bitswap;
 pub use self::block::Block;
 pub use self::error::BitswapError;
-pub use self::ledger::Priority;
+pub use self::ledger::{Priority, Stats};
 pub use self::strategy::{AltruisticStrategy, BitswapStore, BlockPut, Strategy};
 
 mod bitswap_pb {
